+++ app/include/footer.php
@@ -1,7 +1,13 @@
 <footer class="sticky-footer">
 	<div class="container my-auto">
 		<div class="copyright text-center my-auto">
-			<span>alebupal - <a href="https://github.com/alebupal/tvspy" target="_blank">TVSPY 2.0</a> - <a href="https://www.paypal.me/alebupal" target="_blank">PayPal</a> - <a href="https://hub.docker.com/r/alebupal/tvspy/" target="_blank">Docker</a></span>
+			<span>
+				alebupal -
+				<a href="https://github.com/alebupal/tvspy" target="_blank">TVSPY 2.0<span class="nuevaVersion"></span></a> -
+				<a href="https://github.com/alebupal/tvspy/blob/master/CHANGELOG.md" target="_blank">Changelog</a> -
+				<a href="https://www.paypal.me/alebupal" target="_blank">PayPal</a> -
+				<a href="https://hub.docker.com/r/alebupal/tvspy/" target="_blank">Docker</a>
+			</span>
 		</div>
 	</div>
 </footer>